[BugFix][DevTool] Fix all debugging capabilities which rely on DevToolPlatformAndroidDelegate being unavailable when rendering template by preloading
When preloading a Lynx page, the LynxView is initially null when creating LynxInspectorOwner. Subsequently, the LynxView will be reattached to LynxInspectorOwner via the attach method. However, a null check for the LynxView was added when constructing DevToolPlatformAndroidDelegate for LynxInspectorOwner.

As a result, if the LynxView is null, the DevToolPlatformAndroidDelegate object is also null and will not be created thereafter. Consequently, all UI-related debugging features that rely on DevToolPlatformAndroidDelegate become unavailable when rendering cards via the preloading method.

To fix this:
- Remove null check for LynxView when creating DevToolPlatformAndroidDelegate.
- Attach LynxView for DevToolPlatformAndroidDelegate when LynxView is ready.

issue:f-6274446107
AutoLand: release/3.1, release/3.0