[BugFix][Infra] Fixed LynxLite compilation issues caused by refactoring and feature enrichment
LynxDevtool depends on ENABLE_JUST_LEPUSNG after refactoring. In order to compile Lynx Lite normally, we add this macro to LynxDevtool.

When enable_lite=true, adjust the parameters in the Lynx.gni file that are associated with the enable_lite setting. Ensure these parameters are correctly configured to align with the enable_lite activation

issue: f-6167069164

AutoLand: release/3.2,release/3.1,release/3.0