[Feature][NativeModule] Support Using NAPI Modules in BTS within Restricted NAPI Environment
In this MR, we use NAPI as the cross-platform module API for Lynx. Users can load and use their custom NAPI modules in the background-thread script. It's important to note that users cannot use `napi_run_script` and `napi_get_global` in NAPI modules, as we provide a restricted NAPI environment to ensure the stability of the script runtime. For detailed usage, please refer to the documentation.

The main changes in this MR are as follows:
1. Abstract an interface class `NapiRuntimeProxyInterface` from `NapiRuntimeProxy`.

2. Add a `RestrictedNapiRuntimeProxyDecorator` to wrap the `NapiRuntimeProxy` object. This decorator will provide a restricted NAPI environment with some hooked APIs externally, and it will inject a restricted NAPI loader into JS at specific times.

3. Add `PrepareRestrictedNapiEnvironment` in `lynx_runtime.cc` and call this function during runtime initialization to create a restricted NAPI environment.

4. Add a `getModuleLoader` interface in `lynx.ts` for users to obtain the restricted NAPI loader.

doc: https://github.com/lynx-family/lynx/discussions/845
AutoSubmit: True
SkipChecks: macro