[BugFix][List] Fix list overflow behavior error.
If list is overflow visible, mView's parent will be set clip children to false in super.layout() which lets mView not be clipped to its bound. Consider list item view's parent is the LinearLayout, so we should also invoke mView.setClipChildren(false) to make LinearLayout not be clipped to its bound.

issue: m-6577225178