[Feature][EmbeddedMode] Introduce EmbeddedMode switch in LynxViewBuilder.
In this MR, we introduced the EmbeddedMode switch, which is passed through LynxViewBuilder and then to the C++ LynxShell layer.

When the switch is enabled, there will be targeted performance optimizations and restrictions on business flexibility.

Specific optimizations will be integrated in future MRs.

It is important to note that this switch is still experimental, so it is not recommended for users to enable it on their own.

doc: https://github.com/lynx-family/lynx/issues/1125

[internal] doc: https://bytedance.larkoffice.com/wiki/FHhfwsSKjiTo0Zkm11sckwy6nkM [end-internal]

issue: m-6577263377

AutoSubmit: true