[BugFix][Android][Event] Fixed the issue with incorrect WindowRect obtained in folding screen mode.
description:
In order to avoid frequent IPC, the exposure was previously obtained from LynxContext ScreenMetrics to calculate WindowRect, but when no external interface is called to update ScreenMetrics (such as when the folding screen is unfolded), WindowRect may be wrong, resulting in abnormal exposure.
Therefore, we obtain ScreenMetrics through DisplayMetricsHolder, but only call it at the appropriate time.

AutoLand: release/3.3, release/3.2

issue: m-5736536125