[Optimize] Add compiling flag for `flex_optional` to always wrap T using memory save template
Previously we introduced `flex_optional` to dynamically choose from `std::optional` and `std::unique_ptr` when T is of different size. But for some fields of `ComputedCSSStyle`, they are seldomly used but still be inlined by `std::optional`. We think the 32-byte threshold is reasonable, so we provide another way to force flex_optional to choose the internal implementation of std::unique_ptr.

issue: m-6383995029