[Optimize] Add benchmark cases for later optimizations introduced to base::Vector
Subsequently, relevant optimizations will be introduced for `base::Vector`, which is expected to significantly improve array expansion, insert, and erase methods.

When submitting the code this time, relevant performance test cases are added to the repository in advance so that real performance data can be obtained when the code is optimized and merged later.

This commit adds more test cases covering `base::String` as value type and `<base::String, lepus::Value>` as map value type.