[BugFix] Fix header file of basedevtool not found
This PR[1] changed the packaging method, which caused the directory structure of the baseDevtool source package to change.
However, the header search paths were not updated accordingly, resulting in header files from BaseDevtool not being found by other modules that depend on it.
Here, additional search paths are added to adapt to this.

[1]: https://github.com/lynx-family/lynx/commit/e3e2745e4d9fcdafe0e59916776eb3762fae0f8e