[BugFix] Fixed the issue where lynxviewclient could not receive InitLynxviewEntry.
This MR is to fixed the issue where lynxviewclient could not receive InitLynxviewEntry.

Delay the sending of the init event to solve the problem of lynxviewclient registration being later than the event sending. The fix is to delay the init event until after main thread engine and background thread engine are ready.

AutoSubmit:true
AutoLand:release/3.4
issue: m-6736507039